pyproject.toml
[project]
name = "llpm"
version = "0.1.0"
description = "Markdown-based project management CLI for LLM multi-agent workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
llpm = "llpm.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llpm"]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
